Update radio and check to use new Scarpe/Lacci structure
