+++ lacci/lib/shoes/widgets.rb
@@ -18,11 +18,13 @@ require "shoes/widgets/star"
 
 require "shoes/widgets/alert"
 require "shoes/widgets/button"
+require "shoes/widgets/check"
 require "shoes/widgets/edit_box"
 require "shoes/widgets/edit_line"
 require "shoes/widgets/image"
 require "shoes/widgets/link"
 require "shoes/widgets/list_box"
 require "shoes/widgets/para"
+require "shoes/widgets/radio"
 require "shoes/widgets/span"
 require "shoes/widgets/video"
